Refactor: break up azure_devops.py into multiple modules.
