+++ components/collector/src/source_collectors/api_source_collectors/azure_devops/__init__.py
@@ -1,0 +1,5 @@
+"""Azure Devops Server metric collectors.
+
+Where possible we use version 4.1 of the API so we can support TFS 2018 and newer.
+See https://docs.microsoft.com/en-gb/rest/api/azure/devops/?view=azure-devops-rest-4.1#api-and-tfs-version-mapping
+"""
